+++ modules/controllers/07-secrets.tf
@@ -1,7 +1,5 @@
-resource "aws_secretsmanager_secret" "token" {
-  name = "kubernetes-controller-${var.environment}-${var.kubernetes_cluster}-${random_string.seed.result}-token"
-}
-
-resource "aws_secretsmanager_secret" "ca_hash" {
-  name = "kubernetes-controller-${var.environment}-${var.kubernetes_cluster}-${random_string.seed.result}-hash"
+# Will store a fresh token and the CA hash - known only at runtime
+# TODO: Remove controller
+resource "aws_secretsmanager_secret" "secrets" {
+  name = "kubernetes-controller-${var.environment}-${var.kubernetes_cluster}-${random_string.seed.result}-secrets"
 }