+++ modules/nodes/00-variables_required_inputs.tf
@@ -1,28 +1,3 @@
-//variable "network_region" {
-//  type        = string
-//  description = "The AWS region where to spin the infrastructure"
-//}
-
-//variable "unique_identifier" {
-//  type        = string
-//  description = "A UUID to be used in resource names generation to avoid conflicts"
-//}
-
-//variable "environment" {
-//  type        = string
-//  description = "The environment to use"
-//}
-
-//variable "controller_join_token" {
-//  type        = string
-//  description = "kubeadm control join token. This needs to be unique for each cluster"
-//}
-
-//variable "kubernetes_cluster" {
-//  type        = string
-//  description = "Cluster name indentifier"
-//}
-
 variable "vpc_id" {
   type        = string
   description = "The VPC id"
@@ -38,23 +13,22 @@ variable "k8s_workers_num_nodes" {
   description = "Number of nodes for the asg for the nodes"
 }
 
-//variable "internal_network_cidr" {
-//  type        = string
-//  description = "TEMPORARY: Allow access from a certain ip range" # TODO: FIXME: This needs to be removed and the sg exported
-//}
-
 variable "nodes_cri_bootstrap" {
   type = string
 }
 
-variable "private_subnets" {
-  //  type = list(string)
-}
-
-//variable "controller_sg_id" {
-//  type = string
-//}
+variable "private_subnets" {}
 
 variable "nodes_config_bundle" {
-
+  type = object({
+    environment             = string
+    k8s_deb_package_version = string
+    kubernetes_cluster      = string
+    controller_join_token   = string
+    unique_identifier       = string
+    controllers_sg_id       = string
+    internal_network_cidr   = string
+    secret_name             = string
+    secret_arn              = string
+  })
 }