+++ modules/nodes/03-data_sources.tf
@@ -8,14 +8,10 @@ data "aws_vpc" "targeted_vpc" {
 
 data "template_file" "bootstrap_k8s_controllers_kubeadm_join_config" {
   template = file("${path.module}/scripts/kubeadm_join_config.yaml")
-  vars = {
-    controller_join_token = local.controller_join_token
-  }
 }
 
 data "template_file" "bootstrap_node_k8s_workers" {
   template = file("${path.module}/scripts/bootstrap.sh")
-
   vars = {
     controller_join_token   = local.controller_join_token
     cluster_id              = local.kubernetes_cluster
@@ -24,5 +20,6 @@ data "template_file" "bootstrap_node_k8s_workers" {
     post_install            = var.userdata_post_install
     kubeadm_join_config     = data.template_file.bootstrap_k8s_controllers_kubeadm_join_config.rendered
     cri_installation        = var.nodes_cri_bootstrap
+    secret_name             = local.secret_name
   }
 }