+++ modules/nodes/scripts/bootstrap.sh
@@ -14,6 +14,7 @@ KCTL_USER='ubuntu'
 echo "START: $(date)" >>/opt/bootstrap_times
 
 AWS_HOSTNAME=$(curl http://169.254.169.254/latest/meta-data/local-hostname)
+
 hostname "$AWS_HOSTNAME"
 echo "$AWS_HOSTNAME" >/etc/hostname
 echo "127.0.0.1 $AWS_HOSTNAME" >>/etc/hosts
@@ -110,6 +111,32 @@ EOF
 # Replacing with the master ip
 sed -i "s/MASTERIP/$MASTER_IP/g" "/home/$KCTL_USER/kubeadm-join-config.yaml"
 
+#=======================================================================================================================
+# Get a fresh join token and the CA Hash
+
+# TODO: Need to wait for the token to be there
+# TODO: FINISH THIS
+SECRET_NAME=${secret_name}
+while true; do
+	TOKEN=$(aws secretsmanager get-secret-value --secret-id "$SECRET_NAME" --region "$AWS_REGION" | jq --raw-output '.SecretString' | jq --raw-output '.token')
+	HASH=$(aws secretsmanager get-secret-value --secret-id "$SECRET_NAME" --region "$AWS_REGION" | jq --raw-output '.SecretString' | jq --raw-output '.hash')
+	#  echo TOKEN: "$TOKEN"
+	#  echo HASH: "$HASH"
+	# shellcheck disable=SC2000
+	if [[ $(echo "$HASH" | wc -c) == "65" ]]; then
+		echo "Value found..."
+		break
+	else
+		echo "Wait 10 seconds..."
+		sleep 10
+	fi
+done
+# Substitute the join token and hash
+sed -i "s/CONTROLLERJOINTOKEN/$TOKEN/g" "/home/$KCTL_USER/kubeadm-join-config.yaml"
+sed -i "s/CAHASH/$HASH/g" "/home/$KCTL_USER/kubeadm-join-config.yaml"
+
+#=======================================================================================================================
+
 OLD_HOME=$HOME
 export HOME=/root # Fix bug: https://github.com/kubernetes/kubeadm/issues/2361
 kubeadm join --config "/home/$KCTL_USER/kubeadm-join-config.yaml" --v=5
@@ -117,11 +144,11 @@ kubeadm join --config "/home/$KCTL_USER/kubeadm-join-config.yaml" --v=5
 echo 'source /usr/share/bash-completion/bash_completion' >>$HOME/.bashrc
 HOME=$OLD_HOME
 
+# TODO: This file does not exist
 # FIX CIS: [FAIL] 4.2.6 Ensure that the --protect-kernel-defaults argument is set to true (Automated)
 echo 'protectKernelDefaults: true' >>/var/lib/kubelet/config.yaml
 
 # Adding autocomplete
-echo 'source /usr/share/bash-completion/bash_completion' >>/root/.bashrc
 echo 'source <(kubectl completion bash)' >/etc/bash_completion.d/kubectl
 echo 'source <(kubeadm completion bash)' >/etc/bash_completion.d/kubeadm